+++ src/createExtendedRoutine.js
@@ -1,6 +1,6 @@
 import { createRoutineCreator, defaultRoutineStages } from 'redux-saga-routines';
 
-export default function createExtendedRoutine(typePrefix, stages, payloadCreator, metaCreator){
+export default function createExtendedRoutine(typePrefix, stages = [], payloadCreator, metaCreator){
   if(typeof typePrefix !== 'string'){
     throw new Error('`typePrefix` must be a string');
   };